Move observables to their own directory; Start optimizing Observable statistics computations
